move cmd entries to submodule
