+++ sourced/ml/cmd_entries/__init__.py
@@ -1,0 +1,6 @@
+from sourced.ml.cmd_entries.bigartm2asdf import bigartm2asdf_entry
+from sourced.ml.cmd_entries.bow_converters import bow2vw_entry, nbow2bow_entry
+from sourced.ml.cmd_entries.dump_model import dump_model
+from sourced.ml.cmd_entries.postprocess_id2vec import postprocess_id2vec
+from sourced.ml.cmd_entries.projector_entry import projector_entry
+from sourced.ml.cmd_entries.run_swivel import run_swivel
