+++ sourced/ml/cmd_entries/projector_entry.py
