+++ sourced/ml/tests/test_id2vec.py
@@ -7,7 +7,8 @@ import numpy
 
 import sourced.ml.tests.models as paths
 from sourced.ml.models import Id2Vec
-from sourced.ml.utils import projector, projector_entry
+from sourced.ml.utils import projector
+from sourced.ml.cmd_entries import projector_entry
 
 
 class Id2VecTests(unittest.TestCase):
