+++ sourced/ml/utils/__init__.py
@@ -1,5 +1,3 @@
-from sourced.ml.utils.bigartm import bigartm2asdf_entry, install_bigartm
-from sourced.ml.utils.dump import dump_model
+from sourced.ml.utils.bigartm import install_bigartm
 from sourced.ml.utils.engine import create_engine, create_spark
 from sourced.ml.utils.pickleable_logger import PickleableLogger
-from sourced.ml.utils.projector_entry import projector_entry
