+++ sourced/ml/utils/bigartm.py
@@ -4,13 +4,7 @@ import multiprocessing
 import os
 import shutil
 import subprocess
-import sys
 import tempfile
-
-import numpy
-from scipy.sparse import csr_matrix
-
-from sourced.ml.models import Topics
 
 
 def execute(cmd, cwd, log):
@@ -63,38 +57,3 @@ def install_bigartm(args=None, target="./bigartm", tempdir=None, warn_exists=Tru
         shutil.copyfile(os.path.join(cwd, "bin", "bigartm"), target)
         os.chmod(target, 0o777)
     log.info("Installed %s", os.path.abspath(target))
-
-
-def bigartm2asdf_entry(args):
-    """
-    BigARTM "readable" model -> Topics -> Modelforge ASDF.
-    """
-    tokens = []
-    data = []
-    indices = []
-    indptr = [0]
-    if args.input != "-":
-        fin = open(args.input)
-    else:
-        fin = sys.stdin
-    try:
-        # the first line is the header
-        fin.readline()
-        for line in fin:
-            items = line.split(";")
-            tokens.append(items[0])
-            nnz = 0
-            for i, v in enumerate(items[2:]):
-                if v == "0":
-                    continue
-                nnz += 1
-                data.append(float(v))
-                indices.append(i)
-            indptr.append(indptr[-1] + nnz)
-    finally:
-        if args.input != "-":
-            fin.close()
-    data = numpy.array(data, dtype=numpy.float32)
-    indices = numpy.array(indices, dtype=numpy.int32)
-    matrix = csr_matrix((data, indices, indptr), shape=(len(tokens), len(items) - 2)).T
-    Topics().construct(tokens, None, matrix).save(args.output)
